Extract std option to submodule.
